Getting all the movement logic refactored properly.
